Move a conditional test closure to where it needs to occur.
timesSameSymbolWasSeen doesn't need to be tested after it was just
reset back to 1, for example.  We only need to test it when the
current symbol matches the previous one.  In fact, we can even
combine the increment and the test into one statement.

And oh yeah, it's ++timesSameSymbolWasSeen, so that the value is
incremented _before_ it's tested.